+++ src/live_tree.rs
@@ -250,6 +250,19 @@ impl Iter {
                     continue;
                 }
             };
+            if ft.is_dir() {
+                // TODO: Count them?
+                // TODO: Perhaps an option to back them up anyhow?
+                match cachedir::is_tagged(&dir_entry.path()) {
+                    Ok(true) => continue,
+                    Ok(false) => (),
+                    Err(e) => {
+                        ui::problem(&format!(
+                            "Error checking CACHEDIR.TAG in {dir_entry:?}: {e}"
+                        ));
+                    }
+                }
+            }
 
             let metadata = match dir_entry.metadata() {
                 Ok(metadata) => metadata,